add TallestTreeSpecies tests

+++ hadoop-examples-mapreduce/src/main/java/com/opstty/mapper/TallestTreeSpeciesMapper.java
@@ -1,14 +1,15 @@
 package com.opstty.mapper;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
 import java.io.IOException;
 
-public class TallestTreeSpeciesMapper extends Mapper<Object, Text, Text, IntWritable> {
+public class TallestTreeSpeciesMapper extends Mapper<Object, Text, Text, DoubleWritable> {
     private Text word = new Text();
-    private IntWritable height = new IntWritable(1);
+    private DoubleWritable height = new DoubleWritable(1);
     private boolean first = true;
 
     public void map(Object key, Text value, Context context) throws IOException, InterruptedException {
@@ -16,7 +17,7 @@ public class TallestTreeSpeciesMapper extends Mapper<Object, Text, Text, IntWrit
             String[] itr = value.toString().split(";");
             word.set(itr[3]);
             if(!itr[6].isEmpty()) {
-                height.set((int) Double.parseDouble(itr[6]));
+                height.set( Double.parseDouble(itr[6]));
             }
             context.write(word, height);
         } else {
