+++ hadoop-examples-mapreduce/src/main/java/com/opstty/reducer/TallestTreeSpeciesReducer.java
@@ -1,18 +1,19 @@
 package com.opstty.reducer;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 
-public class TallestTreeSpeciesReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
-    private IntWritable result = new IntWritable(1);
+public class TallestTreeSpeciesReducer extends Reducer<Text, DoubleWritable, Text, DoubleWritable> {
+    private DoubleWritable result = new DoubleWritable(1.0);
 
-    public void reduce(Text key, Iterable<IntWritable> values, Context context)
+    public void reduce(Text key, Iterable<DoubleWritable> values, Context context)
             throws IOException, InterruptedException {
-        int height = 0;
-        for (IntWritable val : values) {
+        double height = 0.0;
+        for (DoubleWritable val : values) {
             if(height <= val.get()) {
                 height = val.get();
             }
